start adding pytests
